Implemented UI
Re-arranged the classes to more appropriately handle a UI.
Separated the data in the Key into a separate name and username (name
must be unique, username does not).
Added in documentation for reference.

TODOs to consider: implement the buttons to properly process
adding,editing, and deleting information. Also adjust the Password Class
to take a special class to encrypt and decrypt based on its own
encryption methods, rather than those defined by the main program.